rollback test organization
I was so wrong! The amount of pain this has caused me after 10 minutes of using it... 😅

It's pretty good for running test by groups, but I can't run a specific test in IntelliJ anymore. There must be some conf, to achieve this but... I don't want to deal with this right now